WebExtensions refactor
