+++ components/support/webextensions/src/test/java/mozilla/components/support/webextensions/WebExtensionSupportTest.kt
@@ -452,7 +452,7 @@ class WebExtensionSupportTest {
 
         val delegateCaptor = argumentCaptor<WebExtensionDelegate>()
         WebExtensionSupport.initialize(
-            engine = engine,
+            runtime = engine,
             store = store,
             onUpdatePermissionRequest = { _, _, _, _ ->
                 executed = true
